pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.10",
    "langsmith>=0.4.34",
    "notebook>=7.4.7",
    "python-dotenv>=1.1.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "slowapi>=0.1.9",
    "emoji>=2.14.0",
    "boto3>=1.35.0",
    "pytest-json-report>=1.5.0",
]

[tool.pytest.ini_options]
pythonpath = ["agentic-system", "tests/feature-testing"]
markers = [
    "contract: output contract / structural field-presence tests (fast, deterministic)",
    "llm_feature: tests that invoke the real dialog system LLM (slow, costs API credits)",
    "llm_judge: tests that use an LLM judge to evaluate response quality",
    "simulated: Strategy B tests — full conversation simulation from scratch (slowest)",
]
